Crate ic_config

Re-exports

pub use config::*;

pub use config_parser::*;

pub use config_sample::*;

Modules

artifact_pool

config
This module defines the Config data structure which is the config for the (almost) entire replica and related parsing functions. Each component that has an individual Config should get included here so that it can be parsed.

config_parser

config_sample

consensus

crypto

embedders

execution_environment

firewall

http_handler

logger

message_routing

metrics

nns_registry_replicator

registration

registry_client

scheduler

state_manager

subnet_config
This module implements support for static configurations for components that can be different for different subnet types.