Struct ic_config::config::Config

pub struct Config {
    pub registry_client: RegistryClientConfig,
    pub transport: TransportConfig,
    pub state_manager: StateManagerConfig,
    pub scheduler: SchedulerConfig,
    pub hypervisor: HypervisorConfig,
    pub http_handler: HttpHandlerConfig,
    pub metrics: MetricsConfig,
    pub artifact_pool: ArtifactPoolTomlConfig,
    pub consensus: ConsensusConfig,
    pub crypto: CryptoConfig,
    pub logger: LoggerConfig,
    pub nodemanager_logger: LoggerConfig,
    pub message_routing: MessageRoutingConfig,
    pub malicious_behaviour: MaliciousBehaviour,
    pub firewall: FirewallConfig,
    pub registration: RegistrationConfig,
    pub nns_registry_replicator: NnsRegistryReplicatorConfig,
}

The config struct for the replica. Just consists of Configs for the components.

Fields
registry_client: RegistryClientConfigtransport: TransportConfigstate_manager: StateManagerConfigscheduler: SchedulerConfighypervisor: HypervisorConfighttp_handler: HttpHandlerConfigmetrics: MetricsConfigartifact_pool: ArtifactPoolTomlConfigconsensus: ConsensusConfigcrypto: CryptoConfiglogger: LoggerConfignodemanager_logger: LoggerConfigmessage_routing: MessageRoutingConfigmalicious_behaviour: MaliciousBehaviourfirewall: FirewallConfigregistration: RegistrationConfignns_registry_replicator: NnsRegistryReplicatorConfig
Implementations

impl Config

pub fn new(parent_dir: PathBuf) -> Self

Return a Config with default settings that put all paths under a 'parent_dir', with the given 'subnet_id'.

It is an alternative way to construct a Config than parsing a configuration file.

pub fn run_with_temp_config<T>(run: impl FnOnce(Self) -> T) -> T

Run a function with default Config created with temporary directories that are automatically removed after the function finishes.

pub fn temp_config() -> (Self, TempDir)

Return a function with default Config created with temporary directories that are automatically removed once TempDir is dropped. TempDir should remain in scope for the duration where the Config is being used.

pub fn load_with_default(
source: &ConfigSource,
default: Config
) -> Result<Self, ConfigError>

Load Config from the given 'config_descr' where if a section is omitted, its value is taken from the given 'default'.

pub fn load_with_tmpdir(config_source: ConfigSource, tmpdir: PathBuf) -> Config

Load the Replica config from the given source