Module ic_config::subnet_config

This module implements support for static configurations for components that can be different for different subnet types.

Structs

CowMemoryManagerConfig
The per subnet type configuration for CoW Memory Manager

CyclesAccountManagerConfig

SchedulerConfig
The per subnet type configuration for the scheduler component

SubnetConfig
If a component has at least one static configuration that is different for different subnet types, then it is included in this struct.

SubnetConfigs
A struct that holds the per subnet configuration for all the subnet types on the internet computer.